% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opuy.R
\docType{data}
\name{op_uy}
\alias{op_uy}
\alias{opuy}
\title{opuy}
\format{
data.frame con 4788 filas y 12 columnas
}
\usage{
data(opuy)
}
\description{
Contiene dos indicadores relevantes de opinion publica relativo a las
eleccione en Uruguay para el periodo 1989 - 2020: intencion de voto y evaluacion
de la gestion del presidente. Los datos son agregados y provienen de la divulgacion
de las empresas consultoras que realizan las encuestas.
}
\details{
\describe{
   \item{medicion}{}
   \item{empresa}{}
   \item{tipo_eleccion}{}
   \item{anio_medicion}{}
   \item{eleccion}{}
   \item{fecha}{}
   \item{partido}{}
   \item{sigla}{}
   \item{candidato}{}
   \item{valor}{}

}
}
\keyword{datasets}
