% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumen_opuy.R
\name{resumen_opuy}
\alias{resumen_opuy}
\title{resumen opuy}
\usage{
resumen_opuy()
}
\value{
data.frame que contiene los indicadores que cada consultora realizó por anio.
}
\description{
Resumen de los indicadores que contiene la base de datos por consultora.
}
\examples{
resumen_opuy()
}
